NO JIRA. Package names should be lowercase.
