+++ src/test/scala/nl.knaw.dans.easy.bag2deposit/AppSpec.scala
@@ -16,7 +16,7 @@
 package nl.knaw.dans.easy.bag2deposit
 
 import better.files.File
-import nl.knaw.dans.easy.bag2deposit.Fixture.{ AppConfigSupport, BagIndexSupport, FileSystemSupport }
+import nl.knaw.dans.easy.bag2deposit.fixture.{ AppConfigSupport, BagIndexSupport, FileSystemSupport }
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
